Make formatScope reusable by extracting it
